requirements.txt
asgiref==3.7.2
Django==3.2.20
djangorestframework==3.14.0
sqlparse==0.4.4
